+++ app/Http/Controllers/manager/PathologistController.php
@@ -233,8 +233,54 @@ class PathologistController extends Controller
 
             // Detach from pivot table (organization_pathologist)
             $pathologist->organizations()->detach();
+          $id = decrypt($request->id);
+            $pathologist = Pathologist::findOrFail($id);
+
+            if ($pathologist->photo) {
+                $photoPath = public_path(str_replace(asset('/'), '', $pathologist->photo));
+                if (file_exists($photoPath)) {
+                    unlink($photoPath);
+                }
+            }
+
+            // Delete signature
+            if ($pathologist->signature) {
+                $signaturePath = public_path(str_replace(asset('/'), '', $pathologist->signature));
+                if (file_exists($signaturePath)) {
+                    unlink($signaturePath);
+                }
+            }
+
+            // Delete license
+            if ($pathologist->license) {
+                $licensePath = public_path(str_replace(asset('/'), '', $pathologist->license));
+                if (file_exists($licensePath)) {
+                    unlink($licensePath);
+                }
+            }
+
+            // Detach from pivot table (organization_pathologist)
+            $pathologist->organizations()->detach();
+
+            if (!$pathologist) {
+                return response()->json([
+                    'success' => 0,
+                    'message' => 'pathologist not found',
+                ]);
+            }
+
+            $user = User::findOrFail($pathologist->user_id);
             $pathologist->delete();
 
+            if ($user) {
+                $user->delete();
+            }
+
+
+            return response()->json([
+                'success' => 1,
+                'message' => "Pathologist Delete successfully",
+            ]);
             return response()->json([
                 'success' => 1,
                 'message' => "Pathologist Delete successfully",
